* defs.h (error): Add printf format attribute. * thread-db.c (thread_from_lwp): Fix error format string. * stack.c (parse_frame_specification): Ditto. * cli/cli-decode.c (undef_cmd_error): Ditto. * scm-lang.c (scm_lookup_name): Ditto. * tracepoint.c (trace_error): Ditto. * remote-utils.c (usage): Ditto. * remote.c (compare_sections_command): Ditto. Fix PR gdb/328.

+++ gdb/cli/cli-decode.c
@@ -975,7 +975,7 @@ undef_cmd_error (char *cmdtype, char *q)
 	 cmdtype,
 	 q,
 	 *cmdtype ? " " : "",
-	 strlen (cmdtype) - 1,
+	 (int) strlen (cmdtype) - 1,
 	 cmdtype);
 }
